Actualización del metodo de eliminar un usuario por su id , implementamos las excepciones de una eliminacion exitosa , eliminacion fallida por que no existe en la base de datos , y la ultima que al ingresar un id = 0 me muestra un mensaje que se cancela la eliminacion , tambien se realizaron las pruebas pertinentes para corroborar que funciona todo el codigo

+++ api-upao-transporte/src/main/java/com/upao/transporte/controller/UsuarioController.java
@@ -109,10 +109,17 @@ public class UsuarioController {
 
     @DeleteMapping("eliminarUsuario/{id}")
     public ResponseEntity<String> eliminarUsuario(@PathVariable Long id) {
+        if (id == 0){
+            return ResponseEntity.badRequest().body("Se canceló la eliminación de la cuenta. No se proporcionó un ID válido.");
+        }
+
+        if (!plataformaService.existeUsuario(id)) {
+            return ResponseEntity.badRequest().body("La cuenta que desea eliminar ya no existe.");
+        }
+
         plataformaService.eliminarUsuario(id);
         return ResponseEntity.ok("Usuario eliminado correctamente");
     }
-
     @GetMapping("consulta/{id}")
     public ResponseEntity<Usuario> consultarInformacionUsuario(@PathVariable Long id) {
         Usuario user = plataformaService.consultarUsuarioPorId(id);
